Add better handling for CSRF Token.
