+++ frontend/src/app/page.tsx
@@ -1,4 +1,9 @@
+/// File 1: /frontend/src/app/page.tsx
+
 'use client'
+
+import { useEffect } from 'react'
+import axios from 'axios'
 
 import Section1 from '@/components/home/Section1'
 import Section2 from '@/components/home/Section2'
@@ -7,6 +12,19 @@ import Section4 from '@/components/home/Section4'
 import Section5 from '@/components/home/Section5'
 
 export default function Home() {
+  useEffect(() => {
+    const initializeCsrf = async () => {
+      try {
+        await axios.get('/sanctum/csrf-cookie')
+        console.log('CSRF initialized successfully')
+      } catch (error) {
+        console.error('Failed to initialize CSRF', error)
+      }
+    }
+
+    initializeCsrf()
+  }, [])
+
   return (
     <>
       <Section1 />
